Revert "Feat: elasticsearch 인덱싱 API 구현"
This reverts commit 0f39c5c91316ec7852d1ec4e0603fc7f34ac814a.
