+++ src/main/java/com/kimdabang/kdbserver/product/product/application/ProductService.java
@@ -17,4 +17,7 @@ public interface ProductService {
     List<ProductResponseVo> getProductsByCategory(Long categoryId, int page, int size);
     long getTotalPagesByCategory(Long categoryId, int size);
 
+    // elasticSearch
+    public void indexProducts();
+
 }
