+++ src/main/java/com/kimdabang/kdbserver/product/product/domain/ProductDocument.java
@@ -1,13 +1,10 @@
 package com.kimdabang.kdbserver.product.product.domain;
 
-import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.data.elasticsearch.annotations.Document;
-import org.springframework.data.elasticsearch.annotations.Field;
-import org.springframework.data.elasticsearch.annotations.FieldType;
 
 @Getter
 @Builder
@@ -16,11 +13,8 @@ import org.springframework.data.elasticsearch.annotations.FieldType;
 @Document(indexName = "products")
 public class ProductDocument {
 
-    @Id
     private String id;
-    @Field(type = FieldType.Text)
     private String productName;
-    @Field(type = FieldType.Text)
     private String productDescription;
 
 }
