--- src/main/java/com/kimdabang/kdbserver/product/product/infrastructure/ProductSearchRepository.java
@@ -1,12 +1,0 @@
-package com.kimdabang.kdbserver.product.product.infrastructure;
-
-import com.kimdabang.kdbserver.product.product.domain.ProductDocument;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-import java.util.List;
-
-public interface ProductSearchRepository extends ElasticsearchRepository<ProductDocument, String> {
-
-    List<ProductDocument> findByProductNameContainingOrProductDescriptionContaining(String name, String description, Pageable pageable);
-}
